shine-backend/requirements.txt
flask
flask-cors
mysql-connector-python
werkzeug
python-dotenv
